+++ Carbon.Accounts/Functions/Resolve-CPrincipalName.ps1
@@ -1,24 +1,29 @@
 
-function Resolve-CIdentityName
+function Resolve-CPrincipalName
 {
     <#
     .SYNOPSIS
-    Determines the full, NT identity name for a user or group.
+    Determines the full, NT principal name for a user or group.
 
     .DESCRIPTION
-    `Resolve-CIdentityName` resolves a user/group name into its full, canonical name, used by the operating system. For example, the local Administrators group is actually called BUILTIN\Administrators. With a canonical username, you can unambiguously compare principals on objects that contain user/group information.
+    `Resolve-CPrincipalName` resolves a user/group name into its full, canonical name, used by the operating system. For
+    example, the local Administrators group is actually called BUILTIN\Administrators. With a canonical username, you
+    can unambiguously compare principals on objects that contain user/group information.
 
-    If unable to resolve a name into an identity, `Resolve-CIdentityName` returns nothing.
+    If unable to resolve a name into an principal, `Resolve-CPrincipalName` returns nothing.
 
-    If you want to get full identity information (domain, type, sid, etc.), use `Resolve-CIdentity`.
+    If you want to get full principal information (domain, type, sid, etc.), use `Resolve-CPrincipal`.
 
-    In Carbon 2.0, you can also resolve a SID into its identity name. The `SID` parameter accepts a SID in SDDL form as a `string`, a `System.Security.Principal.SecurityIdentifier` object, or a SID in binary form as an array of bytes. If the SID no longer maps to an active account, you'll get the original SID in SDDL form (as a string) returned to you.
+    In Carbon 2.0, you can also resolve a SID into its principal name. The `SID` parameter accepts a SID in SDDL form as
+    a `string`, a `System.Security.Principal.SecurityIdentifier` object, or a SID in binary form as an array of bytes.
+    If the SID no longer maps to an active account, you'll get the original SID in SDDL form (as a string) returned to
+    you.
 
     .LINK
     ConvertTo-CSecurityIdentifier
 
     .LINK
-    Resolve-CIdentity
+    Resolve-CPrincipal
 
     .LINK
     Test-CIdentity
@@ -33,51 +38,39 @@ function Resolve-CIdentityName
     string
 
     .EXAMPLE
-    Resolve-CIdentityName -Name 'Administrators'
+    Resolve-CPrincipalName -Name 'Administrators'
 
     Returns `BUILTIN\Administrators`, the canonical name for the local Administrators group.
     #>
     [CmdletBinding(DefaultParameterSetName='ByName')]
-    [OutputType([string])]
+    [OutputType([String])]
     param(
-        [Parameter(Mandatory=$true,ParameterSetName='ByName',Position=0)]
-        [string]
         # The name of the identity to return.
-        $Name,
+        [Parameter(Mandatory, ParameterSetName='ByName', Position=0)]
+        [String] $Name,
 
-        [Parameter(Mandatory=$true,ParameterSetName='BySid')]
-        # Get an identity's name from its SID. Accepts a SID in SDDL form as a `string`, a `System.Security.Principal.SecurityIdentifier` object, or a SID in binary form as an array of bytes.
+        # Get an identity's name from its SID. Accepts a SID in SDDL form as a `string`, a
+        # `System.Security.Principal.SecurityIdentifier` object, or a SID in binary form as an array of bytes.
         #
         # This parameter is new in Carbon 2.0.
+        [Parameter(Mandatory, ParameterSetName='BySid')]
         $SID
     )
 
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    if( $PSCmdlet.ParameterSetName -eq 'ByName' )
+    if ($PSCmdlet.ParameterSetName -eq 'ByName')
     {
-        return Resolve-CIdentity -Name $Name -NoWarn -ErrorAction Ignore | Select-Object -ExpandProperty 'FullName'
-    }
-    elseif( $PSCmdlet.ParameterSetName -eq 'BySid' )
-    {
-        $SID = ConvertTo-CSecurityIdentifier -SID $SID -NoWarn
-        if( -not $SID )
-        {
-            return
-        }
-
-        $id = [Carbon.Identity]::FindBySid( $SID )
-        if( $id )
-        {
-            return $id.FullName
-        }
-        else
-        {
-            return $SID.ToString()
-        }
+        return Resolve-CPrincipal -Name $Name -ErrorAction Ignore | Select-Object -ExpandProperty 'FullName'
     }
 
+    $id = Resolve-CPrincipal -Sid $SID -ErrorAction Ignore
+    if ($id)
+    {
+        return $id.FullName
+    }
+
+    return $SID.ToString()
 }