+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 
 ### Migration Instructions
 
+* Replace usages of `Resolve-CIdentity` with `Resolve-CPrincipal`.
+* Replace usages of `Test-CIdentity` with `Test-CPrincipal`.
+* Replace usages of `Resolve-CIdentityName` with `Resolve-CPrincipalName`.
+
 ### Changed
 
 * Renamed `Resolve-CIdentity` to `Resolve-CPrincipal`.
-* Renamed `Resolve-CIdentity` to `Resolve-CPrincipal`.
+* Renamed `Test-CIdentity` to `Test-CPrincipal`.
+* Renamed `Resolve-CIdentityName` to `Resolve-CPrincipalName`.